CustomWave now inherits from Waveform.  There is some weirdness, but overall everything is GRRRRRRRRRRRRRREAT!
git-svn-id: https://projectm.svn.sourceforge.net/svnroot/projectm/trunk@1071 6778bc44-b910-0410-a7a0-be141de4315d

+++ src/projectM-engine/Pipeline.hpp
@@ -17,7 +17,7 @@ public:
 
 	 float videoEchoAlpha;
 	 float videoEchoZoom;
-	 float videoEchoOrientation;
+	 int videoEchoOrientation;
 
 	 std::vector<RenderItem*> drawables;
 	 std::vector<RenderItem*> compositeDrawables;
